Update all collision benchmarks for new nomenclature

+++ benchmarks/collisions/Stopping_power123.py
@@ -65,7 +65,7 @@ for path in ["Stopping_power1","Stopping_power2","Stopping_power3"]:
 	wavelength = 1e-6 # meters
 	c = 3e8
 	
-	times = sim.ParticleBinning(diagNumber=0).getAvailableTimesteps()
+	times = np.double(sim.ParticleBinning(diagNumber=0).getAvailableTimesteps())
 	nx = sim.ParticleBinning(diagNumber=0,timesteps=0).get()["x"].size
 	
 	Ekin = np.zeros((nx,len(times)))
