+++ benchmarks/collisions/conductivity.py
@@ -32,7 +32,7 @@ for path in ["conductivity1","conductivity2","conductivity3"]:
 	coulomb_log          = np.double(sim.namelist.Collisions[0].coulomb_log)
 	dt                   = np.double(sim.namelist.Main.timestep)/(2*np.pi)
 	
-	times = sim.ParticleBinning(diagNumber=0).getAvailableTimesteps()
+	times = np.double(sim.ParticleBinning(diagNumber=0).getAvailableTimesteps())
 	
 	vx_mean = np.zeros((ncases,len(times)))
