+++ benchmarks/collisions/temperature_isotropization.py
@@ -15,7 +15,7 @@ wavelength = 1e-6 # meters
 c = 3e8
 coeff = (2.*np.pi/wavelength)**2*re_*c / (2.*np.sqrt(np.pi))
 
-times = sim.ParticleBinning(diagNumber=0).getAvailableTimesteps()
+times = np.double(sim.ParticleBinning(diagNumber=0).getAvailableTimesteps())
 electrons0 = sim.ParticleBinning(0, sum={"x":"all"}).get()
 vx = electrons0["vx"]
 electrons1 = sim.ParticleBinning(1, sum={"x":"all"}).get()
